+++ sdk/dotnet/StreamAnalytics/V20200301/Inputs/OutputArgs.cs
@@ -37,7 +37,7 @@ namespace Pulumi.AzureNative.StreamAnalytics.V20200301.Inputs
         /// The size window to constrain a Stream Analytics output to.
         /// </summary>
         [Input("sizeWindow")]
-        public Input<double>? SizeWindow { get; set; }
+        public Input<int>? SizeWindow { get; set; }
 
         /// <summary>
         /// The time frame for filtering Stream Analytics job outputs.
