+++ sdk/dotnet/StreamAnalytics/V20200301/Output.cs
@@ -49,7 +49,7 @@ namespace Pulumi.AzureNative.StreamAnalytics.V20200301
         /// The size window to constrain a Stream Analytics output to.
         /// </summary>
         [Output("sizeWindow")]
-        public Output<double?> SizeWindow { get; private set; } = null!;
+        public Output<int?> SizeWindow { get; private set; } = null!;
 
         /// <summary>
         /// The time frame for filtering Stream Analytics job outputs.
@@ -155,7 +155,7 @@ namespace Pulumi.AzureNative.StreamAnalytics.V20200301
         /// The size window to constrain a Stream Analytics output to.
         /// </summary>
         [Input("sizeWindow")]
-        public Input<double>? SizeWindow { get; set; }
+        public Input<int>? SizeWindow { get; set; }
 
         /// <summary>
         /// The time frame for filtering Stream Analytics job outputs.
