+++ sdk/dotnet/StreamAnalytics/V20211001Preview/GetOutput.cs
@@ -112,7 +112,7 @@ namespace Pulumi.AzureNative.StreamAnalytics.V20211001Preview
         /// <summary>
         /// The size window to constrain a Stream Analytics output to.
         /// </summary>
-        public readonly double? SizeWindow;
+        public readonly int? SizeWindow;
         /// <summary>
         /// The time frame for filtering Stream Analytics job outputs.
         /// </summary>
@@ -142,7 +142,7 @@ namespace Pulumi.AzureNative.StreamAnalytics.V20211001Preview
 
             object? serialization,
 
-            double? sizeWindow,
+            int? sizeWindow,
 
             string? timeWindow,
