+++ sdk/dotnet/StreamAnalytics/V20211001Preview/Outputs/OutputResponse.cs
@@ -47,7 +47,7 @@ namespace Pulumi.AzureNative.StreamAnalytics.V20211001Preview.Outputs
         /// <summary>
         /// The size window to constrain a Stream Analytics output to.
         /// </summary>
-        public readonly double? SizeWindow;
+        public readonly int? SizeWindow;
         /// <summary>
         /// The time frame for filtering Stream Analytics job outputs.
         /// </summary>
@@ -77,7 +77,7 @@ namespace Pulumi.AzureNative.StreamAnalytics.V20211001Preview.Outputs
 
             object? serialization,
 
-            double? sizeWindow,
+            int? sizeWindow,
 
             string? timeWindow,
