+++ sdk/python/pulumi_azure_native/streamanalytics/v20200301/outputs.py
@@ -3177,7 +3177,7 @@ class OutputResponse(dict):
                  datasource: Optional[Any] = None,
                  name: Optional[str] = None,
                  serialization: Optional[Any] = None,
-                 size_window: Optional[float] = None,
+                 size_window: Optional[int] = None,
                  time_window: Optional[str] = None):
         """
         An output object, containing all information associated with the named output. All outputs are contained under a streaming job.
@@ -3188,7 +3188,7 @@ class OutputResponse(dict):
         :param Union['AzureDataLakeStoreOutputDataSourceResponse', 'AzureFunctionOutputDataSourceResponse', 'AzureSqlDatabaseOutputDataSourceResponse', 'AzureSynapseOutputDataSourceResponse', 'AzureTableOutputDataSourceResponse', 'BlobOutputDataSourceResponse', 'DocumentDbOutputDataSourceResponse', 'EventHubOutputDataSourceResponse', 'EventHubV2OutputDataSourceResponse', 'GatewayMessageBusOutputDataSourceResponse', 'PowerBIOutputDataSourceResponse', 'ServiceBusQueueOutputDataSourceResponse', 'ServiceBusTopicOutputDataSourceResponse'] datasource: Describes the data source that output will be written to. Required on PUT (CreateOrReplace) requests.
         :param str name: Resource name
         :param Union['AvroSerializationResponse', 'CsvSerializationResponse', 'JsonSerializationResponse', 'ParquetSerializationResponse'] serialization: Describes how data from an input is serialized or how data is serialized when written to an output. Required on PUT (CreateOrReplace) requests.
-        :param float size_window: The size window to constrain a Stream Analytics output to.
+        :param int size_window: The size window to constrain a Stream Analytics output to.
         :param str time_window: The time frame for filtering Stream Analytics job outputs.
         """
         pulumi.set(__self__, "diagnostics", diagnostics)
@@ -3264,7 +3264,7 @@ class OutputResponse(dict):
 
     @property
     @pulumi.getter(name="sizeWindow")
-    def size_window(self) -> Optional[float]:
+    def size_window(self) -> Optional[int]:
         """
         The size window to constrain a Stream Analytics output to.
         """
