+++ sdk/python/pulumi_azure_native/streamanalytics/v20211001preview/outputs.py
@@ -4041,7 +4041,7 @@ class OutputResponse(dict):
                  datasource: Optional[Any] = None,
                  name: Optional[str] = None,
                  serialization: Optional[Any] = None,
-                 size_window: Optional[float] = None,
+                 size_window: Optional[int] = None,
                  time_window: Optional[str] = None,
                  watermark_settings: Optional['outputs.OutputWatermarkPropertiesResponse'] = None):
         """
@@ -4054,7 +4054,7 @@ class OutputResponse(dict):
         :param Union['AzureDataLakeStoreOutputDataSourceResponse', 'AzureFunctionOutputDataSourceResponse', 'AzureSqlDatabaseOutputDataSourceResponse', 'AzureSynapseOutputDataSourceResponse', 'AzureTableOutputDataSourceResponse', 'BlobOutputDataSourceResponse', 'DocumentDbOutputDataSourceResponse', 'EventHubOutputDataSourceResponse', 'EventHubV2OutputDataSourceResponse', 'GatewayMessageBusOutputDataSourceResponse', 'PostgreSQLOutputDataSourceResponse', 'PowerBIOutputDataSourceResponse', 'RawOutputDatasourceResponse', 'ServiceBusQueueOutputDataSourceResponse', 'ServiceBusTopicOutputDataSourceResponse'] datasource: Describes the data source that output will be written to. Required on PUT (CreateOrReplace) requests.
         :param str name: Resource name
         :param Union['AvroSerializationResponse', 'CsvSerializationResponse', 'CustomClrSerializationResponse', 'JsonSerializationResponse', 'ParquetSerializationResponse'] serialization: Describes how data from an input is serialized or how data is serialized when written to an output. Required on PUT (CreateOrReplace) requests.
-        :param float size_window: The size window to constrain a Stream Analytics output to.
+        :param int size_window: The size window to constrain a Stream Analytics output to.
         :param str time_window: The time frame for filtering Stream Analytics job outputs.
         :param 'OutputWatermarkPropertiesResponse' watermark_settings: Settings which determine whether to send watermarks to downstream.
         """
@@ -4142,7 +4142,7 @@ class OutputResponse(dict):
 
     @property
     @pulumi.getter(name="sizeWindow")
-    def size_window(self) -> Optional[float]:
+    def size_window(self) -> Optional[int]:
         """
         The size window to constrain a Stream Analytics output to.
         """
